+++ physeqsolvertableheader.h
@@ -9,12 +9,12 @@ class PhysEqSolverTable;
 class PhysEqSolverTableHeader : public QHeaderView {
     Q_OBJECT
 public:
-    PhysEqSolverTableHeader(QList<int>, QWidget *);
+    PhysEqSolverTableHeader(QList<double>, QWidget *);
 
-    void timeSliceList(const QList<int> lstTimeSlices) { m_lstTimeSlices = lstTimeSlices; }
-    QList<int> timeSliceList() const { return m_lstTimeSlices; }
+    void timeSliceList(const QList<double> lstTimeSlices) { m_lstTimeSlices = lstTimeSlices; }
+    QList<double> timeSliceList() const { return m_lstTimeSlices; }
 signals:
-    void updateTimeSlices(const int, const int);
+    void updateTimeSlices(const int, const double);
 public slots:
     void onShowContextMenu(const QPoint &);
     void insertColumn();
@@ -26,7 +26,7 @@ private:
     QAction *m_pActionInsertColumn;
     QAction *m_pActionRemoveColumn;
     int m_hdrIdx;
-    QList<int> m_lstTimeSlices;
+    QList<double> m_lstTimeSlices;
 };
 
 #endif // PHYSEQSOLVERTABLEHEADER_H
